Combine convenience functions into a single module.
