+++ pcgrandom/test/test_pcg_xsl_rr_v0.py
@@ -18,7 +18,7 @@ Tests for the PCG_XSL_RR_V0 generator.
 import pkgutil
 import unittest
 
-from pcgrandom.pcg_xsl_rr_v0 import PCG_XSL_RR_V0
+from pcgrandom import PCG_XSL_RR_V0
 from pcgrandom.test.test_pcg_common import TestPCGCommon
